refactor(folder-structure): model 폴더
#68
