+++ lib/enums/currency_enums.dart
