+++ lib/enums/transaction_enums.dart
