+++ lib/enums/utxo_enums.dart
