+++ lib/enums/wallet_enums.dart
@@ -1,11 +1,11 @@
+enum WalletType {
+  singleSignature,
+  multiSignature,
+}
+
 enum WalletSyncResult {
   newWalletAdded,
   existingWalletUpdated,
   existingWalletNoUpdate,
   existingName, // fail sync
-}
-
-enum WalletType {
-  singleSignature,
-  multiSignature,
 }
