+++ lib/model/api/request/faucet_request.dart
