+++ lib/model/api/response/faucet_response.dart
