+++ lib/model/api/response/faucet_status_response.dart
