+++ lib/model/app/error/app_error.dart
