+++ lib/model/app/wallet/multisig_signer.dart
