+++ lib/model/app/wallet/multisig_wallet_list_item.dart
