+++ lib/model/app/wallet/singlesig_wallet_list_item.dart
