+++ lib/model/app/wallet/wallet_list_item_base.dart
